Add our logger!
Our logger is able to log to syslog, journald, a file (with checking for file
rotation), and the NT event log.  We also catch unhandled exceptions, so that
we can log them.  We pull level and facility from configuration.

The config code is also changed to load the logger first, and to use it.
As part of that, we clean up how we identify which config files to load.