Move user-add code to add callback
Lots of code moving around here, but that’s mostly all it is, moving
code.

Since the refresh-done and add callbacks both deal with adding users,
the bulk of the code has moved to the add callback, and the
refresh-done callback calls the add callback on each user.  The add
callback now returns the list of groups modified, and the refresh-done
callback uses it to build a set of groups, for reporting.

The add callback also gets a named parameter, so the refresh-done
callback can prevent a flood of “user added to group” messages.  Since
the parameter has a default, it doesn’t conflict with Syncrepl calling
us.

The unique/user/groups attribute names are now moved to be class-level
variables, which are set by the main LDAP code before the thread starts.

So, I guess that’s the add code done!