Make attribute value encoding configurable
LDAP can be very dicey about encodings.  The best place to find out how
an attribute is coded is to check the schema.

Instead, we take the easy way out, and all the user to configure what
the encoding is for each attribute.

This sets up the encoding configuration, defaulting everything to ASCII.