The start of the expander code!
The expander main code is very similar to LDAP main code, except that we're
using subprocesses instead of threads, and we're spawning (and maintaining) a
set number of workers.  We use signals to relay messages as appropriate.

The worker is able to read changes from the database, and delete them, all
transactionally.  If there are no pending changes, then we sleep on a LISTEN
call for 60 seconds, and then repeat.

There's still lots to do, like metrics, and handling problem workers, and
workers need to handle signals properly.  Oh, and we're not actually _doing_
anything on the changes yet!  But still, we have a framework now!