Set explicit naming conventions.
This is a change for Alembic compatibility.

In order for Alembic to track automatically-named stuff (like
indexes), instead of using whatever name the DB auto-generates,
we provide a naming convention for SQLAlchemy to use.

To do this, we have to explicitly instantiate a MetaData, which
we then provide to our declarative base constructor.