Make DB sequences semi-separate objects.
This is a change for Alembic compatibility.

We now explicitly declare our sequences as being separate from
their associated column.  We don't completely separate it, though,
since we're keeping the declaration within the table.

As a side-effect, these changes ensure that Postgres is able to
automatically grab the next ID in the sequence—instead of needing
the client to fetch one first—when doing an INSERT.