Install python-apt from source
python-apt is used by apport, which is part of the default site config.
Rather than mess with site config, let's just bring in python-apt.

NOTE: There is a python-apt on PyPi, but it's out of date, is Python 2
only, and is dangerous to use.

python-distutils-extra is brought in because python-apt needs it.