Have cloud-sql-proxy use metadata env
WGLURP's DB config doesn't currently play well with PostgreSQL sockets,
because we're dealing with psycopg2 and SQLAlchemy URLs.

Our previous config for the Cloud SQL Proxy used FUSE because we didn't
know the project name at the time of service start.  But, now that we
have instance metadata put into an env file, we do!