Convert /mnt/data into a systemd mount
I have to remember, systemd manages mounts now.  Changing /etc/fstab
means you have to do a `systemd daemon-reload`.

The good thing about this is, since /mnt/data is encrypted (and so I
need to do an unlock dance), I can explitly define a system unit
for the /mnt/data mount, and make that unit depend on the oneshot
service that unlocks the drive!